♻️ refactor: #4 캘린더 컴포넌트 분리 및 바깥 영역 클릭시 캘린더 닫히도록 수정
